only kill key listener if running [XWin]

+++ src/niacop/niacop/Native/WindowManagers/XWindowManager.cs
@@ -79,8 +79,10 @@ namespace niacop.Native.WindowManagers {
         }
 
         public void deinitialize() {
-            keyHookProc.Kill();
-            keyHookProc.WaitForExit();
+            if (keyHookProc != null && !keyHookProc.HasExited) {
+                keyHookProc.Kill();
+                keyHookProc.WaitForExit();   
+            }
         }
     }
 }
